// client/src/pages/Agents/AgentsLanding/types.ts
export type ValuesTypes = {
  fullName?: string;
  firstName: string;
  lastName: string;
  state: string;
  licenseNumber: string;
  position: string;
  bio: string;
  avatar: string;
  phoneNumber: string;
  emailAddress: string;
  address: string;
  twitter: string;
  instagram: string;
  linkedIn: string;
  facebook: string;
  telNumber: string;
  password: string;
  confirmPassword: string;
  role: string;
  languages: string[];
  specialties: string[];
};